Rename MdCreator and BcacheCreator step 1
