+++ src/lib/y2storage/proposal.rb
@@ -34,9 +34,9 @@ require "y2storage/proposal/autoinst_devices_planner"
 require "y2storage/proposal/autoinst_space_maker"
 require "y2storage/proposal/autoinst_drives_map"
 require "y2storage/proposal/partition_creator"
-require "y2storage/proposal/bcache_creator"
+require "y2storage/proposal/autoinst_bcache_creator"
 require "y2storage/proposal/lvm_creator"
-require "y2storage/proposal/md_creator"
+require "y2storage/proposal/autoinst_md_creator"
 require "y2storage/proposal/partition_killer"
 require "y2storage/proposal/partitions_distribution_calculator"
 require "y2storage/proposal/phys_vol_calculator"
