+++ src/lib/y2storage/proposal/autoinst_bcache_creator.rb
