+++ src/lib/y2storage/proposal/autoinst_md_creator.rb
