+++ test/y2storage/proposal/autoinst_bcache_creator_test.rb
