+++ test/y2storage/proposal/autoinst_md_creator_test.rb
